functioning loggin function, inquiries pseudo implemented
