+++ src/utilities/axiosWithAuth.js
@@ -1,7 +1,6 @@
 import axios from "axios";
 
-export const axiosWithAuth = () => {
-  const token = localStorage.getItem("token");
+export const axiosWithAuth = (token) => {
   return axios.create({
     headers: {
       Bearer: token,
